requirements/dev.txt
aws-cdk-lib
aws-xray-sdk
basilico
boto3
boto3-stubs[dynamodb]
constructs
invoke
mypy
pytest
pytest-cov
pytest-mock
ruff
simplelens
types-boto3
